requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.5
    # via langchain-community
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via httpx
attrs==26.1.0
    # via aiohttp
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.7
    # via requests
dataclasses-json==0.6.7
    # via langchain-community
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.12.0
    # via neo4j-graphrag
greenlet==3.4.0
    # via sqlalchemy
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   langgraph-sdk
    #   langsmith
    #   ollama
httpx-sse==0.4.3
    # via langchain-community
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
json-repair==0.44.1
    # via neo4j-graphrag
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.1.1
    # via jsonpatch
langchain==1.2.15
    # via -r requirements.in
langchain-classic==1.0.4
    # via
    #   langchain-community
    #   langchain-neo4j
langchain-community==0.4.1
    # via -r requirements.in
langchain-core==1.3.0
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-classic
    #   langchain-community
    #   langchain-neo4j
    #   langchain-ollama
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-neo4j==0.9.0
    # via -r requirements.in
langchain-ollama==1.1.0
    # via -r requirements.in
langchain-text-splitters==1.1.2
    # via langchain-classic
langgraph==1.1.8
    # via
    #   langchain
    #   langchain-neo4j
langgraph-checkpoint==4.0.2
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.10
    # via langgraph
langgraph-sdk==0.3.13
    # via langgraph
langsmith==0.7.32
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
marshmallow==3.26.2
    # via dataclasses-json
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
neo4j==6.1.0
    # via
    #   -r requirements.in
    #   langchain-neo4j
    #   neo4j-graphrag
neo4j-graphrag==1.14.1
    # via langchain-neo4j
numpy==2.4.4
    # via
    #   -r requirements.in
    #   langchain-community
    #   neo4j-graphrag
    #   scipy
ollama==0.6.1
    # via langchain-ollama
orjson==3.11.8
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.2
    # via langgraph-checkpoint
packaging==26.1
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pydantic==2.13.2
    # via
    #   langchain
    #   langchain-classic
    #   langchain-core
    #   langgraph
    #   langsmith
    #   neo4j-graphrag
    #   ollama
    #   pydantic-settings
pydantic-core==2.46.2
    # via pydantic
pydantic-settings==2.13.1
    # via langchain-community
pypdf==6.10.2
    # via neo4j-graphrag
python-dotenv==1.2.2
    # via
    #   -r requirements.in
    #   pydantic-settings
pytz==2026.1.post1
    # via neo4j
pyyaml==6.0.3
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
    #   neo4j-graphrag
regex==2026.4.4
    # via tiktoken
requests==2.33.1
    # via
    #   langchain-classic
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
scipy==1.17.1
    # via neo4j-graphrag
sqlalchemy==2.0.49
    # via
    #   langchain-classic
    #   langchain-community
tenacity==9.1.4
    # via
    #   langchain-community
    #   langchain-core
    #   neo4j-graphrag
tiktoken==0.12.0
    # via -r requirements.in
types-pyyaml==6.0.12.20260408
    # via neo4j-graphrag
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via requests
uuid-utils==0.14.1
    # via
    #   langchain-core
    #   langsmith
xxhash==3.6.0
    # via
    #   langgraph
    #   langsmith
yarl==1.23.0
    # via aiohttp
zstandard==0.25.0
    # via langsmith
